+++ app/src/main/java/com/codesoom/assignment/web/TaskHttpHandler.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment;
+package com.codesoom.assignment.web;
 
 import com.codesoom.assignment.models.Task;
 import com.fasterxml.jackson.core.JsonProcessingException;
